+++ NEWS.md
@@ -23,6 +23,7 @@
 - `$<SpatRaster>` can now be used to get a categorical SpatRaster with a different active category
 - `scale_linear<SpatRaster>` method for linear scaling of cell values between a minumum and maximum value such as 0 and 1
 - `distance` and related methods get argument "method" to choose the distance algorithm for lon/lat data [#1677](https://github.com/rspatial/terra/issues/1677) by Márcia Barbosa
+- `divide<SpatRaster>` and `divide<SpatVector>` methods
 
 
 # version 1.8-5